Update BST.py
Adding a function to delete data. This function start with the value that we want to eliminate, if this node has children we have to move them up, that's why we have the "find_inheritor" function. finally we delete the node by notifying to his father that his left or right son is not part of the tree anymore.